% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lblatent-package.R
\docType{package}
\name{lblatent-package}
\alias{lblatent}
\alias{lblatent-package}
\title{lblatent: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
